SOLR-15019: Add support for getting collection metrics and arbitrary node metrics.

+++ solr/core/src/java/org/apache/solr/cluster/placement/ReplicaMetrics.java
@@ -23,6 +23,7 @@ import java.util.Optional;
  *
  */
 public interface ReplicaMetrics {
-  int getReplicaSizeGB();
+
+  Double getReplicaSizeGB();
   Optional<Object> getReplicaMetric(String metricName);
 }
